chore(deps): consolidate Dependabot updates
Folds 4 of the 5 open Dependabot PRs into one change: vite 7 (with the plugin-react and
vite-plugin-svgr bumps its peer ranges require), jest 30 and eslint-plugin-import.

vite-plugin-svgr 5 removed the ReactComponent named export, so 7 SVG imports move to the
?react suffix; jest 30 removed the toBeCalled* aliases, so 5 assertions are renamed.

eslint-plugin-jest 29 stays open - eslint-config-react-app pins typescript-eslint 5.

Closes #573
