+++ package.json
@@ -40,13 +40,13 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^14.3.1",
     "@testing-library/user-event": "^14.5.2",
-    "@vitejs/plugin-react": "^4.2.1",
-    "babel-jest": "^29.7.0",
+    "@vitejs/plugin-react": "^5.2.0",
+    "babel-jest": "^30.2.0",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-jest": "^27.9.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-prettier": "5.1.3",
@@ -55,15 +55,15 @@
     "graphiql": "^3.2.0",
     "graphql": "^16.8.1",
     "husky": "^9.0.11",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "jest": "^30.2.0",
+    "jest-environment-jsdom": "^30.2.0",
     "lint-staged": "^15.2.2",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "vite": "^4.5.3",
-    "vite-plugin-svgr": "^3.3.0",
+    "vite": "^7.2.2",
+    "vite-plugin-svgr": "^5.2.0",
     "web-vitals": "^3.5.2"
   },
   "lint-staged": {
