+++ src/graphiql-explain/components/Profiler.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { getColorByLimit, getSymbolByLimit } from '../utils'
-import { ReactComponent as ArrowIcon } from '../../icons/arrow.svg'
-import { ReactComponent as ChartIcon } from '../../icons/chart.svg'
+import ArrowIcon from '../../icons/arrow.svg?react'
+import ChartIcon from '../../icons/chart.svg?react'
 import { useProfiler } from '../hooks/useProfiler'
 import styles from './Profiler.module.css'
 import { Modal } from './Modal'
