+++ src/graphiql-explain/pages/Explain.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Tabs } from '../components/Tabs'
 import { useTabs } from '../hooks/useTabs'
 import { tabs } from '../constants/tabs'
-import { ReactComponent as StopwatchIcon } from '../../icons/stopwatch.svg'
+import StopwatchIcon from '../../icons/stopwatch.svg?react'
 import styles from './Explain.module.css'
 
 export function Content() {
